Change questions asked in France flow.
The flow for France is being changed.

Rather than asking:

1) Where are you getting married?
2) Do you want to get married or enter into a PACS?

The second question is being changed to:

2) Is your partner the same sex or opposite sex?

The title template has also had to be changed to reflect the question changes.
The `want_to_get_married?` function queried the result of question
`Do you want to get married or enter into a PACS?`. This question is no longer
being asked, so the title template has been changed to query the result of the
`partner same or opposite sex?` question that is being asked.

The content of the test arfacts has not changed, but the files have been renamed.

France still has two outcomes, but now the path to the outcomes is
different:

/france/marriage is now /france/opposite_sex
/france/pacs is now /france/same_sex

Also, Monaco is now the first country in the list that asks the "marriage or pacs"
question, not France, so that file has been renamed..